requirements.txt
altair==5.5.0
attrs==24.3.0
blinker==1.9.0
cachetools==5.3.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.0
fastapi
google-cloud-storage==2.10.0
gitdb==4.0.12
GitPython==3.1.44
googleapis-common-protos==1.66.0
httplib2==0.22.0
idna==3.10
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jwcrypto==1.5.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
mysql-connector-python
narwhals==1.22.0
numpy==2.2.1
packaging==24.2
pandas==2.2.3
pillow==11.1.0
protobuf==5.29.3
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome>=3.4.3
pydantic
pydantic[email]
pydeck==0.9.1
Pygments==2.19.1
pyparsing==3.2.1
pyrebase4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jwt>=2.0.1
pytz==2024.2
referencing==0.35.1
requests==2.32.3
requests-toolbelt==0.10.1
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
setuptools==70.0.0
six==1.17.0
smmap==5.0.2
streamlit==1.41.1
tenacity==9.0.0
toml==0.10.2
tornado==6.4.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==1.26.20
uvicorn==0.17.4
watchdog==6.0.0